services/project-service/pyproject.toml
[project]
name = "project-service"
version = "0.1.0"
description = "Badgers project and conversation service"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "structlog>=24.1.0",
    "httpx>=0.26.0",
    "aio-pika>=9.0.0",
    "python-multipart>=0.0.6",
    "pgvector>=0.2.0",
    "minio>=7.2.0",
    "openai>=1.0.0",
    "PyJWT>=2.9.0",
    "tiktoken>=0.5.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "aiosqlite>=0.20.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"
